+++ src/main/java/thebetweenlands/client/event/ClientRegistrationEvents.java
@@ -297,6 +297,7 @@ public class ClientRegistrationEvents {
 		event.registerEntityRenderer(EntityRegistry.FRESHWATER_URCHIN.get(), FreshwaterUrchinRenderer::new);
 		event.registerEntityRenderer(EntityRegistry.CAVE_FISH.get(), CaveFishRenderer::new);
 		event.registerEntityRenderer(EntityRegistry.DARK_DRUID.get(), DarkDruidRenderer::new);
+		event.registerEntityRenderer(EntityRegistry.FROG.get(), FrogRenderer::new);
 
 		event.registerBlockEntityRenderer(BlockEntityRegistry.MUD_BRICK_ALCOVE.get(), AlcoveRenderer::new);
 		event.registerBlockEntityRenderer(BlockEntityRegistry.ALEMBIC.get(), AlembicRenderer::new);
@@ -381,6 +382,7 @@ public class ClientRegistrationEvents {
 		event.registerLayerDefinition(BLModelLayers.FRESHWATER_URCHIN, FreshwaterUrchinModel::create);
 		event.registerLayerDefinition(BLModelLayers.CAVE_FISH, CaveFishModel::create);
 		event.registerLayerDefinition(BLModelLayers.DARK_DRUID, DarkDruidModel::create);
+		event.registerLayerDefinition(BLModelLayers.FROG, FrogModel::create);
 
 		event.registerLayerDefinition(BLModelLayers.CORRUPT_GECKO, CagedGeckoModel::createCorruptGecko);
 		event.registerLayerDefinition(BLModelLayers.GECKO, GeckoModel::create);