+++ src/main/java/thebetweenlands/common/datagen/BLLanguageProvider.java
@@ -1025,6 +1025,7 @@ public class BLLanguageProvider extends LanguageProvider {
 		this.addEntityAndEgg(EntityRegistry.FRESHWATER_URCHIN, "Freshwater Urchin");
 		this.addEntityAndEgg(EntityRegistry.CAVE_FISH, "Cave Fish");
 		this.addEntityAndEgg(EntityRegistry.DARK_DRUID, "Dark Druid");
+		this.addEntityAndEgg(EntityRegistry.FROG, "Frog");
 
 		this.addEntityType(EntityRegistry.ANGRY_PEBBLE, "Thrown Angry Pebble");
 		this.addEntityType(EntityRegistry.BETWEENSTONE_PEBBLE, "Betweenstone Pebble");