+++ src/main/java/thebetweenlands/common/datagen/BLRegistryProvider.java
@@ -28,7 +28,8 @@ public class BLRegistryProvider extends DatapackBuiltinEntriesProvider {
 		.add(BLRegistries.Keys.ASPECT_ITEMS, AspectItemRegistry::bootstrap)
 		.add(BLRegistries.Keys.ASPECT_TYPES, AspectTypeRegistry::bootstrap)
 		.add(BLRegistries.Keys.ELIXIR_RECIPES, ElixirRecipeRegistry::bootstrap)
-		.add(Registries.DAMAGE_TYPE, DamageTypeRegistry::bootstrap);
+		.add(Registries.DAMAGE_TYPE, DamageTypeRegistry::bootstrap)
+		.add(BLRegistries.Keys.FROG_VARIANT, FrogVariantRegistry::bootstrap);
 
 	public BLRegistryProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> provider) {
 		super(output, provider, BUILDER, Set.of("minecraft", TheBetweenlands.ID));